Day 15
For reasons I cannot work out, for part 2 of the problem, the equivalent
Java implementation only takes ~15s whereas the Rust implementation
takes ~75s. I expected the Rust version to be faster because this time
around, I trimmed the history to only retain the last two times a number
was spoken.