Improve Day 15 Performance
To improve performance, I used a more space-efficient algorithm that
only needs to store the last time a number was spoken. In addition, I
learned that by default, HashMap uses a cryptographically-secure hash
function. So I plugged in the faster, non-cryptographically-secure Fx
Hash instead.

Performance is better at ~22s. However, it's still not as fast as the
~15s sub-optimal algorithm implemented in Java.