Enable unit tests against _public_ samples.
This uses the sample input provided in the problems for the unit tests.
Generally, the same sample is used for both part 1 and part 2. In some
cases, the sample for part 2 is different and not revealed until the
completion of part 1. In addition, the inputs were modified to be usable
by both parts. So this _may_ lead to leakage about part 2 for anyone who
has not completed part 1. Proceed at your own risk. If you prefer to
avoid spoilers, refer only to the data available for 2020:
https://adventofcode.com/2020/ .

To test against your own input, create a directory in the project root
with the same file names used in the `sample` directory. Then, create a
file called `config.toml` with a single key `input_directory` that
points to the one you created. The location is assumed to be relative to
the project root.